+++ cluster/cloudcfg/cloudcfg_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestCloudConfig(t *testing.T) {
 	cfgTemplate = "({{.QuiltImage}}) ({{.SSHKeys}}) ({{.UbuntuVersion}}) " +
-		"({{.MinionOpts}}) ({{.LogLevel}})"
+		"({{.MinionOpts}}) ({{.LogLevel}}) ({{.DockerOpts}})"
 
 	log.SetLevel(log.InfoLevel)
 	ver = "master"
@@ -18,7 +18,7 @@ func TestCloudConfig(t *testing.T) {
 		SSHKeys:    []string{"a", "b"},
 		MinionOpts: MinionOptions{Role: db.Master},
 	})
-	exp := "(quilt/quilt:master) (a\nb) (xenial) (--role \"Master\") (info)"
+	exp := "(quilt/quilt:master) (a\nb) (xenial) (--role \"Master\") (info) ()"
 	if res != exp {
 		t.Errorf("res: %s\nexp: %s", res, exp)
 	}
@@ -27,9 +27,10 @@ func TestCloudConfig(t *testing.T) {
 	ver = "1.2.3"
 	res = Ubuntu(Options{
 		SSHKeys:    []string{"a", "b"},
-		MinionOpts: MinionOptions{Role: db.Worker},
+		MinionOpts: MinionOptions{Role: db.Worker, TLSDir: "dir"},
 	})
-	exp = "(quilt/quilt:1.2.3) (a\nb) (xenial) (--role \"Worker\") (debug)"
+	exp = "(quilt/quilt:1.2.3) (a\nb) (xenial) (--role \"Worker\" " +
+		"--tls-dir \"dir\") (debug) (-v dir:dir:ro)"
 	if res != exp {
 		t.Errorf("res: %s\nexp: %s", res, exp)
 	}