+++ quiltctl/command/connection.go
@@ -5,24 +5,34 @@ import (
 
 	"github.com/quilt/quilt/api"
 	"github.com/quilt/quilt/api/client"
-	"github.com/quilt/quilt/connection/credentials"
+	"github.com/quilt/quilt/connection"
+	"github.com/quilt/quilt/quiltctl/command/credentials"
 )
 
 type connectionFlags struct {
-	host string
+	host   string
+	tlsDir string
 }
 
 func (cf *connectionFlags) InstallFlags(flags *flag.FlagSet) {
 	flags.StringVar(&cf.host, "H", api.DefaultSocket, "the host to connect to")
+	flags.StringVar(&cf.tlsDir, "tls-dir", "",
+		"the directory in which to lookup tls certs")
 }
 
 type connectionHelper struct {
+	creds  connection.Credentials
 	client client.Client
 
 	connectionFlags
 }
 
-func (ch *connectionHelper) BeforeRun() error {
+func (ch *connectionHelper) BeforeRun() (err error) {
+	// Load the credentials that will be used by Quilt clients and servers.
+	ch.creds, err = credentials.Read(ch.tlsDir)
+	if err != nil {
+		return err
+	}
 	return ch.setupClient(client.New)
 }
 
@@ -31,6 +41,6 @@ func (ch *connectionHelper) AfterRun() error {
 }
 
 func (ch *connectionHelper) setupClient(getter client.Getter) (err error) {
-	ch.client, err = getter(ch.host, credentials.Insecure{})
+	ch.client, err = getter(ch.host, ch.creds)
 	return err
 }